// src/enums/EnumsByRules.ts
export enum EnumRuleFormatting {
    includedata = 'Содержит данные',
    noincludedata = 'Не содержит данных',
    textinclude = 'Текст содержит',
    textnoinclude = 'Текст не содержит',
    textstartswith = 'Текст начинается с',
    textendswith = 'Текст заканчивается на',
    textaccurate = 'Текст в точности',

    numberequalto = 'Равно',
    numbernotequalto = 'Не равно',
    numbermorethan = 'Больше',
    numbermorethanorequalto = 'Больше или равно',
    numberlessthan = 'Меньше',
    numberlessthanorequalto = 'Меньше или равно',
    numberbetween = 'Между',
    numbernobetween = 'Не между',

    dateequalto = 'Дата',
    datebefore = 'Дата до',
    dateafter = 'Дата после',

    /* dateinbetw = 'Дата в промежутке',
    dateoutbetw = 'Дата не в промежутке' */
}